pyproject.toml
[tool.poetry]
name = "document_analyzer"
version = "0.1.0"
description = "A comprehensive ML and LLM utility library"
authors = ["manuel <manuelpazpintor@gmail.com>"]
readme = "README.md"
packages = [{include = "document_analyzer"}]

[tool.poetry.dependencies]
cryptography = "3.1"
python = "3.11.10"
numpy = ">=1.24.0"
pandas = ">=2.0.0"
scikit-learn = ">=1.3.0"
mlflow = ">=2.8.0"
langchain = ">=0.1.0"
langchain-google-genai = ">=0.0.5"
python-dotenv = ">=1.0.0"
google-generativeai = ">=0.3.0"
transformers = ">=4.36.0"
torch = ">=2.1.0"
sentence-transformers = ">=2.2.0"
fastapi = ">=0.104.0"
onnxruntime = "1.22.0"
langchain-google-vertexai = "2.0.24"
chromadb = "1.0.12"
ipykernel = "^6.29.5"
beautifulsoup4  = "4.13.4"
webdriver-manager = "4.0.2"
selenium = "4.33.0"
langchain-community = "0.3.25"
pypdf = "5.6.0"
mcp = {extras = ["cli"], version = "^1.9.4"}
typer = "^0.16.0"
[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
ruff = "^0.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ml_utils --cov-report=term-missing"